CORE-5114 PR suggestion: Extract KryoTestUtils companing methods to standalone functions
