+++ libs/ledger/ledger-common-impl/src/test/kotlin/net/corda/ledger/common/impl/transaction/serializer/PrivacySaltImplKryoSerializerTest.kt
@@ -1,6 +1,6 @@
 package net.corda.ledger.common.impl.transaction.serializer
 
-import net.corda.kryoserialization.testkit.KryoTestUtils
+import net.corda.kryoserialization.testkit.createCheckpointSerializer
 import net.corda.ledger.common.impl.transaction.PrivacySaltImpl
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -10,7 +10,7 @@ class PrivacySaltImplKryoSerializerTest {
     fun `serialization of a PrivacySalt`() {
         val privacySalt = PrivacySaltImpl("1".repeat(32).toByteArray())
 
-        val serializer = KryoTestUtils.createCheckpointSerializer(
+        val serializer = createCheckpointSerializer(
             mapOf(
                 PrivacySaltImpl::class.java to PrivacySaltImplKryoSerializer()
             )
