+++ libs/serialization/serialization-kryo/src/test/kotlin/net/corda/kryoserialization/serializers/ClassSerializerTest.kt
@@ -3,7 +3,7 @@ package net.corda.kryoserialization.serializers
 import com.esotericsoftware.kryo.Kryo
 import com.esotericsoftware.kryo.io.Input
 import com.esotericsoftware.kryo.io.Output
-import net.corda.kryoserialization.testkit.KryoTestUtils.Companion.mockSandboxGroup
+import net.corda.kryoserialization.testkit.mockSandboxGroup
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
